Integrated frontend Dashboard consumed calories with backend. Updated some mock data. Frontend now works with reworked database

+++ Database/DTO/FoodRecord/FoodRecordDtoAggregate.cs
@@ -4,7 +4,7 @@ namespace FitnessApp.Database.DTO.FoodRecord
 {
     public class FoodRecordDtoAggregate
     {
-        public int Calories { get; set; }
+        public float Calories { get; set; }
         public float Proteins { get; set; }
         public float Carbohydrates { get; set; }
         public float Fats { get; set; }
